requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0
matplotlib>=3.7.0
plotly>=5.15.0
tensorboard>=2.13.0

# Music processing
music21>=9.1.0
mido>=1.2.10
pretty_midi>=0.2.10

# Machine learning
transformers>=4.30.0
scikit-learn>=1.3.0

# Evaluation metrics
mir-eval>=0.7

# Visualization and demo
streamlit>=1.25.0
gradio>=3.40.0

# Development tools
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pre-commit>=3.3.0

# Optional dependencies
wandb>=0.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
